Add irq to data_queue
Adding an irq might be needed, since that could be handled by the
generic code alone. Deleting the job-enque interface, since that
might be problematic. The current idea how to implement the job-manager
imposes that the kernel really comes to a stop to signal job completion.
if this is not the case, like in the fixed kernel assignment mode,
it will not easily work out and is more complicated than assumed.
That migth not justify any effort in this.

Signed-off-by: Frank Haverkamp <haver@linux.vnet.ibm.com>