Add first example of homemade action
Now we need to fill in the functionality and mimic the action
starting and stopping. This needs to be used in the action
to start the action functionality. Maybe some helper functions
are required to get this nicely working in a transparent fashion,
such that writing software modles hopefully becomes quite simple.

Signed-off-by: Frank Haverkamp <haver@linux.vnet.ibm.com>